requirements.txt
PyYAML==6.0.1
sentence-transformers==3.1.1
ruff==0.5.7
